+++ app/client/src/actions/evaluationActions.ts
@@ -2,8 +2,6 @@ import type { ReduxAction } from "./ReduxActionTypes";
 import { ReduxActionTypes } from "ee/constants/ReduxActionConstants";
 import { intersection } from "lodash";
 import type { DependencyMap } from "utils/DynamicBindingUtils";
-import type { QueryActionConfig } from "entities/Action";
-import type { DatasourceConfiguration } from "entities/Datasource";
 import type { DiffWithNewTreeState } from "workers/Evaluation/helpers";
 import {
   EVALUATE_REDUX_ACTIONS,
@@ -78,46 +76,6 @@ export const setDependencyMap = (
   return {
     type: ReduxActionTypes.SET_EVALUATION_INVERSE_DEPENDENCY_MAP,
     payload: { inverseDependencyMap },
-  };
-};
-
-// Called when a form is being setup, for setting up the base condition evaluations for the form
-export const initFormEvaluations = (
-  // TODO: Fix this the next time the file is edited
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  editorConfig: any,
-  // TODO: Fix this the next time the file is edited
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  settingConfig: any,
-  formId: string,
-) => {
-  return {
-    type: ReduxActionTypes.INIT_FORM_EVALUATION,
-    payload: { editorConfig, settingConfig, formId },
-  };
-};
-
-// Called when there is change in the data of the form, re evaluates the whole form
-export const startFormEvaluations = (
-  formId: string,
-  formData: QueryActionConfig,
-  datasourceId: string,
-  pluginId: string,
-  actionDiffPath?: string,
-  hasRouteChanged?: boolean,
-  datasourceConfiguration?: DatasourceConfiguration,
-) => {
-  return {
-    type: ReduxActionTypes.RUN_FORM_EVALUATION,
-    payload: {
-      formId,
-      actionConfiguration: formData,
-      datasourceId,
-      pluginId,
-      actionDiffPath,
-      hasRouteChanged,
-      datasourceConfiguration,
-    },
   };
 };
