+++ app/client/src/sagas/ReplaySaga.ts
@@ -76,12 +76,13 @@ import {
 import { AppThemingMode } from "selectors/appThemingSelectors";
 import { generateAutoHeightLayoutTreeAction } from "actions/autoHeightActions";
 import { SelectionRequestType } from "sagas/WidgetSelectUtils";
-import { startFormEvaluations } from "actions/evaluationActions";
+import { startFormEvaluations } from "actions/formEvaluationActions";
 import { getUIComponent } from "pages/Editor/QueryEditor/helpers";
 import { type Plugin, UIComponentTypes } from "entities/Plugin";
 import { getCurrentEnvironmentId } from "ee/selectors/environmentSelectors";
 import { updateAndSaveAnvilLayout } from "layoutSystems/anvil/utils/anvilChecksUtils";
 import type { ReplayOperation } from "entities/Replay/ReplayEntity/ReplayOperations";
+import { objectKeys } from "@appsmith/utils";
 
 export interface UndoRedoPayload {
   operation: ReplayOperation;
@@ -103,7 +104,7 @@ export default function* undoRedoListenerSaga() {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function* openPropertyPaneSaga(replay: any) {
   try {
-    const replayWidgetId = Object.keys(replay.widgets)[0];
+    const replayWidgetId = objectKeys(replay.widgets)[0] as string;
 
     if (!replayWidgetId || !replay.widgets[replayWidgetId].propertyUpdates)
       return;
@@ -155,9 +156,9 @@ export function* postUndoRedoSaga(replay: any) {
       processUndoRedoToasts(replay.toasts);
     }
 
-    if (!replay.widgets || Object.keys(replay.widgets).length <= 0) return;
+    if (!replay.widgets || objectKeys(replay.widgets).length <= 0) return;
 
-    const widgetIds = Object.keys(replay.widgets);
+    const widgetIds = objectKeys(replay.widgets) as string[];
 
     yield put(selectWidgetInitAction(SelectionRequestType.Multiple, widgetIds));
     scrollWidgetIntoView(widgetIds[0]);
@@ -357,6 +358,7 @@ function* replayActionSaga(
                 replayEntity.actionConfiguration,
                 replayEntity.datasource.id || "",
                 replayEntity.pluginId,
+                replayEntity.contextType,
                 u.modifiedProperty,
                 true,
                 datasource?.datasourceStorages[currentEnvironment]
